pyproject.toml
[project]
name = "yandexassistent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipython>=9.5.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "telebot>=0.0.5",
    "yandex-cloud-ml-sdk>=0.15.0",
]
